% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_input.R
\name{select_input}
\alias{select_input}
\title{React Select}
\usage{
select_input(inputId, label, choices, selected = NULL,
  placeholder = "Select...", multi = FALSE, searchable = FALSE,
  clearable = TRUE, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}

\item{placeholder}{The text displayed when no option is selected.}

\item{multi}{Allow the user to select multiple values.}

\item{searchable}{Allow the user to search for matching options}

\item{clearable}{Add a button to remove selected choice(s).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'}.}
}
\description{
A flexible and beautiful Select Input control.
}
\examples{
if (interactive()) {

  library(shiny)
  library(shinyReactWidgets)

  ui <- fluidPage(
    tags$h1("react-select example"),

    select_input(
      inputId = "select",
      choices = month.name,
      multi = FALSE
    ),

    verbatimTextOutput("res")
  )

  server <- function(input, output, session) {

    output$res <- renderPrint({
      input$select
    })

  }

  shinyApp(ui, server)
}
}
