% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_input.R
\name{select_input}
\alias{select_input}
\title{React Select}
\usage{
select_input(inputId, choices, selected = NULL,
  placeholder = "Select...", multi = FALSE, searchable = FALSE,
  clearable = TRUE)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}

\item{placeholder}{The text displayed when no option is selected.}

\item{multi}{Allow the user to select multiple values.}

\item{searchable}{Allow the user to search for matching options}

\item{clearable}{Add a button to remove selected choice(s).}
}
\description{
A flexible and beautiful Select Input control.
}
